feat(api): aircraft api is working now

+++ src/main/kotlin/org/oldjopa/hls/service/AircraftService.kt
@@ -62,7 +62,7 @@ class AircraftService(
     }
 
     fun createAircraftType(req: CreateAircraftEquipmentRequest): Long {
-        val engine = engineRepository.findById(req.engineId.toLong())
+        val engine = engineRepository.findById(req.engineId)
             .orElseThrow { NotFoundException("Engine type ${req.engineCount} not found") }
         val equipment = AircraftEquipment(
             manufacturer = req.manufacturer,
